Fix `deque<T>` to work with fancy pointers.
The only thing that needs to happen here is a bit of const-correctness
on the const iterators. A non-const `deque<T>` has a two-level `__map_`
whose pointers look like this:

    ptr<ptr<T>>

And when we const-iterate this thing, we need to return references to
`const T`. We can safely create the type

    ptr<const ptr<T>>

from `ptr<ptr<T>>` --- although even bothering with that is silly IMHO.
However, the old code incorrectly attempts to create the type

    ptr<const ptr<const T>>

which is a violation of pointer semantics: there is no requirement for
`ptr<T>` and `ptr<const T>` to be interchangeable types, nor even for them
to have the same layout, nor even for them to have the same sizeof!
So we should avoid creating the invalid type.

I think this change will be ABI-non-breaking in practice.