Make promise/future shared states allocator-aware.
They already remember if they were allocated with an `allocator_arg`,
and if so, remember the value of their allocator. With this patch,
they *also* remember the value of the allocated pointer, so that they
can deallocate themselves when their refcount drops to zero.

Before this patch, `__on_zero_shared()` was assuming that `pointer_to`
returned a deallocatable pointer. This is now fixed.

The suspicious raw destructor call (instead of calling
`allocator_traits<_Al>::destroy`) seems to be intentional; N4659 says
"The second constructor uses the allocator `a` to allocate memory
for the shared state," but doesn't say whether `construct` and `destroy`
should be used.