Add a new test case for packaged_task.
The existing "allocator_arg" test case is pretty silly because
it doesn't actually use the allocator for anything! The given type
`A` is small enough that Small Object Optimization kicks in and
so no allocation (or deallocation) takes place.

The new test case doesn't have this issue; it uses the allocator
for real.