Make packaged_task states allocator-aware.
They get most if it for free, but they also need to use the
provided allocator to allocate and deallocate their type-erasure
state. I think this is okay (even though std::function allocator-awareness
was removed in C++17) because packaged_task is move-only whereas
the problem with std::function involved copy-construction.