Add a trivial (yet conforming) implementation of *_pool_resource.
This implementation just goes to the upstream resource for every single
allocation. Also, it stores the allocations in a singly linked list
(*also* allocated via the upstream resource) so that it can deallocate
them later. It would be more efficient to store the allocations in
some kind of vector-ish structure.

Nevertheless, this implementation is technically conforming to the
standard.