Implement my proposed re-fix for LWG 2969.
My impression at the moment is that LWG 2969's fix to
http://eel.is/c++draft/mem.poly.allocator.mem#10.2
was insufficiently similar to the current wording in
http://eel.is/c++draft/allocator.adaptor.members#9.2

The latter talks about constructibility from `inner_allocator_type&`;
the former (after LWG 2969) talks about constructibility from
`polymorphic_allocator` with no ref-qualification. But since we are
perfect-forwarding `*this` through a tuple of references, we definitely
*are* going to try to construct the target object from a `polymorphic_allocator&`
and not from a `polymorphic_allocator`. I believe that the wording in
http://eel.is/c++draft/mem.poly.allocator.mem#10.2
needs to be updated to make the program ill-formed in cases where
that construction is going to fail.

I believe this patch is an appropriate reflection of the improved
wording I suggest for LWG 2969.