<experimental/memory_resource>: Implement monotonic_buffer_resource.
https://reviews.llvm.org/D47111

Repeatedly calling allocate() and release() in a loop should not blow up.
I originally thought this was required by the Standard, and maybe it is,
but that was before I discovered this terrible effect, and before Pablo Halpern
told me that it was not his actual intent.

Since the only purpose of storing `__next_buffer_size_` in the object
was to implement the "looping blows up" misfeature, we can now get rid
of that field, which shrinks `monotonic_buffer_resource` from 56 bytes
down to 48 bytes, matching Boost.Container's implementation.

The misfeature is filed as https://cplusplus.github.io/LWG/issue3120