Implement <memory_resource> as a copy of <experimental/memory_resource>.
I wonder if `memory_resource` needs a key function for its vtable;
consider the precedent of D27387. But out-of-lining a trivial destructor
just to avoid `-Wweak-vtables` feels kind of awful.

Experimental and std differ in their implementation of `options()`
because the latter needs to be compileable as `-std=c++11` whereas
the former gets to use `-std=c++14`.

https://reviews.llvm.org/D47360