Trivially relocatable (1/3): Traits and algorithms.
This implements the library behavior proposed in P1144R3.
This patch doesn't do any special optimizations (e.g. to `std::vector`);
those will come later. This is just the bare minimum to conform to
P1144's additions to the standard.

We mangle all the P1144 names behind `__libcpp_...`, so that they
aren't taken by the user-programmer to be official Standard C++ names.