Optimize polymorphic_allocator by warranting the triviality of its construct and destroy methods.
Consider:

    template<class T>
    void foo(std::pmr::vector<T>& v)
    { v.reserve(10); }

Before this patch:

    foo<std::pmr::vector<int>> 745 instructions
    foo<std::vector<int>>      129 instructions
    foo<std::unique_ptr<int>>  169 instructions
    foo<std::pair<int,int>>    103 instructions
    foo<int>                   172 instructions (yes, this makes no sense)

After this patch:

    foo<std::pmr::vector<int>> 745 instructions
    foo<std::vector<int>>       71 instructions
    foo<std::unique_ptr<int>>   58 instructions
    foo<std::pair<int,int>>     58 instructions
    foo<int>                    58 instructions

The reason `std::vector<int>` takes 71 when everyone else takes 58
seems to be entirely due to the fact that `sizeof(vector)` is 24 bytes,
so some loop-stride calculation has to do a division by 3 instead of
a right-shift. Crazy, but true.

I have no idea why `int` is so much worse than `std::pair<int,int>`
in the "before" case.