P1144R3: Implement relocate_at and update uninitialized_relocate{,_n}.
Since P1144R1, uninitialized_relocate has been normatively equivalent
to "move in a loop, then destroy in a loop," to avoid the possibility
of destroying half of the input range and then throwing an exception.
Rewriting into "move-and-destroy in a loop" is an optimization that
we opt into only if the things being moved are nothrow-move-constructible.

P1144R3 (post-Kona) will introduce `relocate_at`, which is the single-object
form of `uninitialized_relocate`.